+++ src/components/crowd-voice/AnalysisSection.tsx
